filter/src/requirements.txt
opencv-python>=4.8.0
numpy>=1.26.0
matplotlib>=3.8.0
scikit-image>=0.22.0
Pillow>=10.0.0
scipy>=1.11.0
pytorch-msssim>=1.0.0
PyWavelets>=1.4.0 